heterodyne: Fix FWHM ipv peak for 5-pt pointing
Previously the gaussian fit constrained the peak to be the highest
value in each cut through the 5-point pattern.  It seems unlikely
that the peak would actually always be one of the three values in
each cut.  This returns the fitting to be more like the situation
of commit 45cc8e5e40 before it was changed in commit 7832c288fd.

This requires converting the FWHM (the calculation of which was
left in the primitive even after it stopped being used) to pixels.
Since the WCS is already obtained as a Starlink::AST object, this
can be done via its "Rate" method.

Extract the cuts into two different files so that they can be inspected
afterwards.  Also construct the names of these files normally rather
than using ORAC::TempFile to make this more convenient.

The general parameters for fitgauss are now only listed once and then
used for the fits in both directions.  The "varuse" parameter is now
set false to disable usage of the variance and give the three points
equal weight.