heterodyne: Display fitted gaussian for pointing
For 5-point pointing observations, use evalfit to generate "model"
files which can be displayed using the "datamodel" method.

The NDF sections used to extract the horizontal and vertical cuts
through the 5-point pattern were adjusted to include extra (blank)
pixels to the sides, so that the output models generated by evalfit
extend beyond the pointing pattern.  The input file to evalfit is
also expanded with sqorst so generate more points for a better plot
of the fitted gaussian.