JSA: do not force product on tag for preview only
When _TAG_AS_REPRESENTATIVE_ is called with the PREVIEW_ONLY option,
instead of calling "force_product_update", simply call the "product"
method in order for this information to be available when we later
call "product_id".  This is because we presumably don't want to
update the headers of the input file if PREVIEW_ONLY was specified.
It also avoid header syncing issues if the spectral WCS is missing.