heterodyne: Display fitted polynomial for focus
Now after fitting the focus position with fitpoly, use evalfit
to generate a model file.  This is then plotted with the data
using the datamodel plot type.  (Note that to do this we need
to delay displaying the data from the collapse primitives to
the _CALCULATE_FOCUS_ primitive so that the fit parameters
are available.)